now at 4.00.00 in tutorial

+++ routes/web.php
@@ -35,34 +35,34 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, "index"]);
 
 // show create form
-Route::get('/listings/create',  [ListingController::class, "create"]);
+Route::get('/listings/create',  [ListingController::class, "create"])->middleware("auth");
 
 // store listing data
-Route::post('/listings/',  [ListingController::class, "store"]);
+Route::post('/listings/',  [ListingController::class, "store"])->middleware("auth");
 
 // show edit form
-Route::get('/listings/{listing}/edit',  [ListingController::class, "edit"]);
+Route::get('/listings/{listing}/edit',  [ListingController::class, "edit"])->middleware("auth");
 
 // update listing
-Route::put('/listings/{listing}', [ListingController::class, "update"]);
+Route::put('/listings/{listing}', [ListingController::class, "update"])->middleware("auth");
 
 // delete listing
-Route::delete('/listings/{listing}', [ListingController::class, "destroy"]);
+Route::delete('/listings/{listing}', [ListingController::class, "destroy"])->middleware("auth");
 
 // single listing
 Route::get('/listings/{listing}',  [ListingController::class, "show"]);
 
 // show register/create form
-Route::get('/register',  [UserController::class, "create"]);
+Route::get('/register',  [UserController::class, "create"])->middleware("guest");
 
 // create new user
 Route::post('/users',  [UserController::class, "store"]);
 
 // log user out
-Route::post("/logout", [UserController::class, "logout"]);
+Route::post("/logout", [UserController::class, "logout"])->middleware("auth");
 
 // show login form
-Route::get("/login", [UserController::class, "login"]);
+Route::get("/login", [UserController::class, "login"])->name("login")->middleware("auth");
 
 // log in user
 Route::post("/users/authenticate", [UserController::class, "authenticate"]);
